[Configs] Small cleanups to the config code.
