+++ config/Cargo.toml
@@ -35,6 +35,7 @@ url = { workspace = true }
 
 [dev-dependencies]
 aptos-crypto = { workspace = true }
+aptos-types = { workspace = true, features = ["fuzzing"] }
 
 [features]
 default = []
